% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsbuild.R
\name{tsbuild}
\alias{tsbuild}
\title{Build a time-series object from case-based data.}
\usage{
tsbuild(df, TSstart, TSstop = TSstart, value,
  TSformat = "YYYY-MM-DD HH:MM:SS")
}
\arguments{
\item{df}{A data frame.}

\item{TSstart}{Time stamp.}

\item{TSstop}{Time stamp.}

\item{value}{Value to aggregate.}

\item{TSformat}{A string containing the TS format. Defaults to "YYYY-MM-DD HH:MM:SS"}
}
\value{
A time-series object based on \code{df}.
}
\description{
Build a time-series object from case-based data.
}
\examples{
tsbuild(df)
}
